requirements.txt
qdrant-client==1.7.0
cohere==4.37
python-dotenv==1.0.0
